Getting verify tests to run with scatter

+++ test/onnx/verify/softmaxcrossentropyloss_no_attr.cpp
@@ -34,7 +34,7 @@ TEST_CASE(softmaxcrossentropyloss_2d_no_reduction_test)
     migraphx::shape score_shape{migraphx::shape::float_type, {4, 4}};
     std::vector<float> score_data = {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1};
     migraphx::shape label_shape{migraphx::shape::int32_type, {4}};
-    std::vector<float> label_data = {0, 3, 1, 2};
+    std::vector<int32_t> label_data = {0, 3, 1, 2};
 
     migraphx::parameter_map pp;
     pp["0"] = migraphx::argument(score_shape, score_data.data());
@@ -56,16 +56,16 @@ TEST_CASE(softmaxcrossentropyloss_2d_sum_reduction_test)
     migraphx::shape score_shape{migraphx::shape::float_type, {4, 4}};
     std::vector<float> score_data = {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1};
     migraphx::shape label_shape{migraphx::shape::int32_type, {4}};
-    std::vector<float> label_data = {0, 3, 1, 2};
+    std::vector<int32_t> label_data = {0, 3, 1, 2};
 
     migraphx::parameter_map pp;
     pp["0"] = migraphx::argument(score_shape, score_data.data());
     pp["1"] = migraphx::argument(label_shape, label_data.data());
 
     auto result = p.eval(pp).back();
-    std::vector<int32_t> result_vector;
+    std::vector<float> result_vector;
     result.visit([&](auto output) { result_vector.assign(output.begin(), output.end()); });
-    std::vector<int32_t> gold = {45730, 44641, 46108, 45010, 46486, 45379, 46864, 45748};
+    std::vector<float> gold = {45730, 44641, 46108, 45010, 46486, 45379, 46864, 45748};
     EXPECT(migraphx::verify::verify_rms_range(result_vector, gold));
 }
 
@@ -78,7 +78,7 @@ TEST_CASE(softmaxcrossentropyloss_2d_mean_reduction_test)
     migraphx::shape score_shape{migraphx::shape::float_type, {4, 4}};
     std::vector<float> score_data = {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1};
     migraphx::shape label_shape{migraphx::shape::int32_type, {4}};
-    std::vector<float> label_data = {0, 3, 1, 2};
+    std::vector<int32_t> label_data = {0, 3, 1, 2};
 
     migraphx::parameter_map pp;
     pp["0"] = migraphx::argument(score_shape, score_data.data());
